Create Fetch Thread Abstraction
